langchain/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs @ file:///croot/attrs_1734533101012/work
backoff==2.2.1
backports.tarfile @ file:///home/conda/feedstock_root/build_artifacts/backports.tarfile_1733325779670/work
bcrypt==4.2.1
beautifulsoup4==4.12.3
Brotli @ file:///croot/brotli-split_1736182456865/work
build @ file:///home/conda/feedstock_root/build_artifacts/python-build_1733230610871/work
CacheControl @ file:///home/conda/feedstock_root/build_artifacts/cachecontrol-split_1736337859595/work
cachetools==5.5.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
cfgv==3.4.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo @ file:///croot/cleo_1705431334181/work
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
crashtest @ file:///croot/crashtest_1679422372509/work
cryptography @ file:///croot/cryptography_1732129644106/work
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
dulwich @ file:///home/conda/feedstock_root/build_artifacts/dulwich_1734640057777/work
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.7
fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1733235979760/work/dist
filelock @ file:///croot/filelock_1700591183607/work
filetype==1.2.0
firebase-admin==6.6.0
firestore==0.0.8
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
identify==2.6.3
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///croot/importlib_metadata-suite_1732633488278/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
iniconfig==2.0.0
installer @ file:///home/conda/feedstock_root/build_artifacts/python-installer_1733237321392/work
instructor==1.7.0
iso8601==2.1.0
jaraco.classes @ file:///tmp/build/80754af9/jaraco.classes_1620983179379/work
jaraco.context @ file:///home/conda/feedstock_root/build_artifacts/jaraco.context_1733382590553/work
jaraco.functools @ file:///home/conda/feedstock_root/build_artifacts/jaraco.functools_1733746366381/work
jeepney @ file:///tmp/build/80754af9/jeepney_1627537048313/work
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
keyring @ file:///home/conda/feedstock_root/build_artifacts/keyring_1735210185992/work
kubernetes==32.0.0
langchain==0.3.16
langchain-anthropic==0.3.4
langchain-community==0.3.16
langchain-core==0.3.32
langchain-google-firestore==0.5.0
langchain-google-genai==2.0.9
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.3.2
lockfile==0.12.2
markdown-it-py==3.0.0
marshmallow==3.26.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mmh3==5.1.0
monotonic==1.6
more-itertools @ file:///croot/more-itertools_1727185441804/work
mpmath==1.3.0
msgpack @ file:///opt/conda/conda-bld/msgpack-python_1652362659880/work
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging @ file:///croot/packaging_1734472117206/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pkginfo @ file:///home/conda/feedstock_root/build_artifacts/pkginfo_1733734533957/work
platformdirs==4.3.6
pluggy==1.5.0
poetry @ file:///home/conda/feedstock_root/build_artifacts/poetry_1736760786401/work
poetry-core @ file:///home/conda/feedstock_root/build_artifacts/poetry-core_1736675068308/work
poetry-plugin-export @ file:///home/conda/feedstock_root/build_artifacts/poetry-plugin-export_1736761824921/work
posthog==3.11.0
pre_commit==4.0.1
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL @ file:///croot/pyopenssl_1723651579422/work
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks @ file:///croot/pyproject_hooks_1679584411881/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rapidfuzz @ file:///croot/rapidfuzz_1700060069902/work
referencing @ file:///croot/referencing_1699012038513/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
requests-toolbelt @ file:///home/conda/feedstock_root/build_artifacts/requests-toolbelt_1733734787568/work
rich==13.9.4
rpds-py @ file:///croot/rpds-py_1736541261634/work
rsa==4.9
SecretStorage @ file:///croot/secretstorage_1678709481048/work
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.45.3
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit @ file:///croot/tomlkit_1728650307440/work
tqdm==4.67.1
trove-classifiers @ file:///croot/trove-classifiers_1729277230900/work
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1718978550903/work
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp @ file:///croot/zipp_1732630741423/work
zstandard==0.23.0
